.github/scripts/requirements.txt
requests>=2.31.0
python-frontmatter>=1.0.0
gitpython>=3.1.40